atlas: always draw cursors and selection with contrasting colors
The atlas engine brought over some of the DX engine's "workaround
choices" -- namely, drawing colored cursors behind text and drawing
selections by alpha blending -- which were made due to limitations in
how we approached the text rendering problem.

Atlas, however, makes *correct* cursor and selection color
implementations trivial, even in the face of ligatures (yes: the cursor
doesn't need to break a ligature to render part of it in a different
color!)

This commit introduces cursor and selection foregrounds. For now,
they're automatically calculated based on the lightness of the
background color: dark backgrounds get white foregrounds, and light
backgrounds get black ones.

We could have done all the lightness detection in the pixel shader, but
since we do eventually want to support customizable foregrounds (as well
as revert and invert!) the infrastructure here is designed to support
them.